requirements.txt
beautifulsoup4==4.8.2
certifi==2020.6.20
chardet==3.0.4
deezer-python==1.5.0
et-xmlfile==1.0.1
fuzzywuzzy==0.18.0
idna==2.10
jdcal==1.4.1
numpy==1.18.1
openpyxl==3.0.4
pandas==1.1.0
python-dateutil==2.8.1
python-Levenshtein-wheels==0.13.1
pytz==2019.3
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
urllib3==1.25.10
xlrd==1.2.0
